requirements.txt
colorama
python-dotenv
web3
numpy